+++ Sources/JiraAPI/Models/Response/CloudResourceResponse.swift
@@ -10,8 +10,12 @@ import Foundation
 public extension JiraAPI.Models {
     struct CloudResourceResponse: Decodable {
         let id: JiraAPI.Auth.CloudID
-        let name: String
-        let scopes: [JiraAPI.Auth.Scope]
-        let avatarUrl: Foundation.URL
+        public let name: String
+        public let url: URL
+        public let scopes: [JiraAPI.Auth.Scope]
+        let avatarUrl: URL
+        
+        public var cloudID: JiraAPI.Auth.CloudID { id }
+        public var avatar: URL { avatarUrl }
     }
 }