+++ Tests/JiraAPITests/ModelResponseTests.swift
@@ -1,6 +1,6 @@
 //
 //  ModelResponseTests.swift
-//  
+//
 //
 //  Created by Andy on 3.4.2021.
 //
@@ -37,6 +37,33 @@ final class ModelResponseTests: XCTestCase {
 
         XCTAssertEqual(object.token, "ACCESS_TOKEN", "should decode token properly")
         XCTAssertNil(object.refresh, "refresh token should be nil")
+    }
+
+    func testCloudResourceResponse() throws {
+        let data = """
+        [
+          {
+            "id": "1324a887-45db-1bf4-1e99-ef0ff456d421",
+            "name": "Site name",
+            "url": "https://your-domain.atlassian.net",
+            "scopes": [
+              "write:jira-work",
+              "read:jira-user",
+              "manage:jira-configuration"
+            ],
+            "avatarUrl": "https://site-admin-avatar-cdn.prod.public.atl-paas.net/avatars/240/flag.png"
+          }
+        ]
+        """
+        .data(using: .utf8)!
+
+        let object = try JSONDecoder().decode([JiraAPI.Models.CloudResourceResponse].self, from: data).first!
+
+        XCTAssertEqual(object.cloudID, "1324a887-45db-1bf4-1e99-ef0ff456d421", "should decode response properly")
+        XCTAssertEqual(object.name, "Site name", "should decode response properly")
+        XCTAssertEqual(object.url, URL(string: "https://your-domain.atlassian.net")!, "should decode response properly")
+        XCTAssertEqual(object.scopes, ["write:jira-work", "read:jira-user", "manage:jira-configuration"])
+        XCTAssertEqual(object.avatar, URL(string: "https://site-admin-avatar-cdn.prod.public.atl-paas.net/avatars/240/flag.png")!)
     }
 
     static var allTests = [